Перенес функции выбора похожих фильмов на уровень выше - в App
