+++ src/components/app/app.jsx
@@ -3,7 +3,30 @@ import {Switch, Route, BrowserRouter} from "react-router-dom";
 import MovieDetails from "../movie-details/movie-details.jsx";
 import PropTypes from "prop-types";
 import Main from "../main/main.jsx";
+import {SHOWING_MOVIES_COUNT} from "../../utils/consts.js";
 
+const getUniqueArrayElements = (array) => {
+  let result = [];
+
+  array.forEach((item) => {
+    if (!result.includes(item)) {
+      result.push(item);
+    }
+  });
+
+  return result;
+};
+
+const getFilmsByFilter = (array, filterType) => array.filter((item) => item.genres.find((it) => it === filterType));
+
+const getSimilarFilmsByGenre = (moviesArray, movie) => {
+  const filmsByGenre = [];
+  const similarFilms = movie.genres.map((genre) => [].concat(getFilmsByFilter(moviesArray, genre)));
+  similarFilms.forEach((movieElement) => movieElement.forEach((it) => filmsByGenre.push(it)));
+  const similarFilmsByGenre = filmsByGenre.filter((it) => it !== movie);
+
+  return getUniqueArrayElements(similarFilmsByGenre);
+};
 
 class App extends PureComponent {
   constructor(props) {
@@ -39,9 +62,12 @@ class App extends PureComponent {
       );
     }
 
+    const similarMovies = getSimilarFilmsByGenre(movies, movies[movieIndex]).slice(0, SHOWING_MOVIES_COUNT);
+
     return (
       <MovieDetails
         movie={movies[movieIndex]}
+        similarMovies={similarMovies}
         onMovieCardClick={this._handleMovieCardClick}
       />
     );
